Update Java
Update minimally for now, just to get index servers up and running again.

For this, our fork of ansible-role-java was updated to 9.3.0.
Adoptium has replaced AdoptOpenJDK. In the version 9.3.0 Adoptium
support was added. But also Centos7 support dropped. We'll try
this verison anyway. Otherwise we would have to update Centos7's
replacement.